requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.4.0
